api/requirements.txt
Flask==1.1.2
SQLAlchemy==1.3.19
Flask-SQLAlchemy==2.4.4
Flask-Script==2.0.6
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
click==7.1.2
psycopg2==2.8.6
gunicorn==20.0.4
pytest==6.1.1
bcrypt==3.2.0
PyJWT==1.7.1
requests==2.24.0